+++ src/plugin-api/main/uk/co/reecedunn/intellij/plugin/processor/debug/DebugSession.kt
@@ -33,6 +33,8 @@ interface DebugSession {
 
     fun stepOver()
 
+    fun stepOut()
+
     val stackFrames: List<XStackFrame>
 
     val suspendContext: XSuspendContext